Use update_datestore=False in save.
Change f3573b017a658c4a00a2befd6b224c006bf9a537 re-added a save call to
the activate_user method that was removed by the TLDAP change.
Unfortunately it did not specify the update_datestore=False flag, so it
tried to update the ldap data which hasn't yet been created at this
point in the code.

This resulted in the tests failing with the following error:
tldap.exceptions.DoesNotExist: person matching query does not exist.