Fix groups.
Store groups in Django so that we don't rely on retrieving group
assignments from LDAP.

Has significant schema changes. In particular we remove the gid
fields from Institute and softwarePackage, and the users fields
from Projects. These fields are replaced with a ForeignKey
reference to Group, which is a new generic model that represents
groups.

Have added the appropriate function to datastores to support this
new group type using the AccountsDataStore. Groups are created
for every machine category configured.

Currently the group GUI doesn't support adding/removing members.
It should be possible to add this later on.

Change-Id: I089f6ade4e39c0b059e83129b889820f57980519