Fix issues creating objects.
If pk was set on object creation, would be treated as update not a
create. Futhermore some datastore methods not correctly called.

Simplify this code, instead of trying to distinguish create vs.  updates
which both do the same thing anyway, combine both into a save operation
which will check which operation needs to be done.

Also update comments in delete() methods.

Change-Id: Ief16836287f856efb8687fe8b766a0feadcb221f