Redo institute models move app.
Django got confused and though people.institute was an old content type
after the people app migration, and prompted the user to delete them.

This is because the contenttypes don't get renamed until the institutes
migration, which hasn't happened yet.

Move all the migrations to the people model, and have the institutes
migration depend on the people migration.

Change-Id: I62e90252e20874a4436fe7353228bb9e07381898