Don't autoapprove if ADMIN_APPROVE_ACCOUNTS set.
Always require administrator to approve application if
ADMIN_APPROVE_ACCOUNTS is set, even if not creating accounts or
projects.

This is because extra work manual work may be required to add user to
project. Alternatively datastores may need to be updated, which are
only configured on the kgadmin not kgreg server.

Change-Id: I9831d79320f55a24890ebce6c6bb2ed7bc78d313