Change the way we manage permissions.
Currently we cannot distinguish if the user is looking at the admin
site or the registration site except via separate URL schemes.

This change is the first in a series that will make it possible
to move registration and admin site differences to karaage-admin
and karaage-register.

If the user visits karaage-admin, access will be denied unless the
user as an administrator.

If the user visits karaage-register, administration rights will be
ignored (accept for a limited set of operations).

Change-Id: Ic7e1bac2d8e425b37adb92c921607bc4ad859c91