Change can_view parameter.
Change to user to request parameter.

This makes it clearer that it is designed to test the currently logged
in user.

More importantly, it makes it possible for can_view() to call is_admin()
to correctly assert if the user has administation rights.

Currently admnistrators have no administration rights if accessing the
registration site, as this is intended for end users. As a result
can_view() will not return the correct results for this case.

Change-Id: I17b3441293c887a36399e4cf64cdef145e4a75aa