Check if userapplication content type exists.
Django won't create the userapplication content type on new installs,
possibly because it no longer exists anymore.

This is despite the fact it did exist when
karaage/applications/migrations/0001_initial.py was created, and this
migration should be creating the content type, too.

So we shouldn't have heart failure because we can't find it.

Change-Id: Ib8ee81b41759a29597a4ffd69a23d5a7b67165a3