Remove change_default_project xmlrpc function.
This function was completely broken anyway, as it operates on an
account, but the arguments don't provide enough information to uniquely
identify an acccount.

Fixing it would require adding extra arguments. Removing instead as it
is not clear if this function is even used.

This function can be added again at a later time if it was really
required.

Change-Id: I0644df2f63e943589511479b1bf3cacd36454c25