Remove redundant code.
We created InstituteQuota and ProjectQuotas, but did not save them.

This is rather significant, as it means neither
machine_category_institute_save() or
machine_category_machine_category_project_save() are called at the
datastore.

Possibly what we are doing should be considered an error condition.  We
are joining an account for a machine category to a project or institute
that has no knowledge of this machine category. So the datastore will
receive notification that the project or institute was joined, but no
notification that the project or institute was created in the first
place. We do not currently check for this at present.

Change-Id: I4eae4d60b37ecfbd9c69de7c6254f948799317b1