Fix datastore issue.
Details as https://github.com/Karaage-Cluster/karaage/issues/234.

We need to filter these events so we only call add_account_to_project
and add_account_to_institute for project/institute that are in the
relevant machine categories. Or rather, we should only call
add_account_to_project and add_account_to_institute if both the account
and the institute belong to the same machine category.

This is a work in progress, we need to update the tests now to work
correctly.

Change-Id: Ia27537141822eedfcb545295629a39d4113ef539