Honour 'make_leader' for application approval
Currently the make_leader value on an Application is ignored when
a new project is created (user is always made a leader) and also
when a user is added to an existing project (user is never made a
leader).

This changes ProjectApplication.approve() to only make a user a leader
when Application.make_leader == True.

For backwards compatibility, I've changed the new_application() view
to explicitly set make_leader=True for a new project application.

Change-Id: I2244eeb46ed18c1b366139e3d6c244c5e19ecf19