Add "make leader" field to project select form.
This form field, Make leader, was always suppose to exist, however
was accidentally deleted.

As such, it was still possible to set it by manually injecting POST
data.

If this field is desirable or not is another question. It is possible
project leaders may not notice the applicant has set this field when
approving applications.

Change-Id: I5a3d9204b2be90553c6304784df266753fc779ae