Django 1.6: Ensure TLDAP gets configured.
TLDAP is configured by having "tldap.django" in the INSTALLED_APPS.

With Django 1.6, INSTALLED_APPS only get configured when something
accesses the database.

If there is a session defined, then when SamlUserMiddleware calls
request.user.get_user(), it will query the database to get session
information. This will configure the INSTALLED_APPS, even if the user is
not logged in.

If however there is no session defined, then when SamlUserMiddleware
calls request.user.get_user(), it will not query the database, and
INSTALLED_APPS will not be configured when control reaches
TransactionMiddleware.

This results in the following error:

Internal Server Error: /karaage/
Traceback (most recent call last):
  File "/usr/lib/python2.7/dist-packages/django/core/handlers/base.py", line 200, in get_response
    response = middleware_method(request, response)
  File "/usr/lib/python2.7/dist-packages/tldap/middleware.py", line 40, in process_response
    if tldap.transaction.is_managed():
  File "/usr/lib/python2.7/dist-packages/tldap/transaction.py", line 105, in is_managed
    for using in tldap.connections:
TypeError: 'NoneType' object is not iterable

This is not an issue with Django 1.6, as everything gets configured
before middleware is called.

Change-Id: Ie2ea678aaea3888c4b21224ed96c67410d61a830