Adapt to schema change
Previously, a person could be a leader of a project without being a
member of that project. This meant other members could not view the
leader unless they were a member of the project.

Now it is not possible to be a leader of a project without being a
member of the project, meaning all users can now view the leader.

We also have increased the number of members in TestProject1 by 1.