Change groups to point to a parent
Groups here are replicated to clusters as posix groups (/etc/groups)
with their associated members.  Groups should relate to a parent which
provides the motivation for having the group, for example members of a
project should be members of the group associated with the project - the
parent generic foreign key in this instance would refer to the project.

The parent relationship also helps to avoid shadow namespacing issues
wherethings like project PIDs and institute names are the same (think
"RMIT"). The application code is forced to resolve any naming conflicts,
and group membership synchronisation issues are no longer possible.

Change-Id: I2822ccca4dd9b837f7a3ec861fc02982b9c45c9f