Simplify autoconfiguration of plugins.
When scanning for urls and optional template files, only look at
Karaage Plugins.

This means that we don't automatically incorporate urls from every
third party application we use.

It also means we can require the existance of urls.py for every plugin,
and import errors in urls.py for plugins will not get silently ignored.

Karaage itself is not considered a plugin, which in turn means we
should merge karaage/urls to karaage/conf/urls to simplify locating
this information.

Change-Id: Ic6d64c4246cb1a84a6d895d9bd9104a787026509