Remove is_authenticated() method
Django 1.10 changes from a function to a method. The fact we override it
with a function breaks backwards compatability.

There is no need for us to override it, as the default is to always
return True anyway. So we remove this function.

Django 2.0 will require all calls be changed to use a property, not s
function.

Change-Id: I1898eaddd91c0f904df6ed062e58d82490a16d85