Ensure applications can conflict with each other
Reaffirm decision made in 8fd31744cd215cc8937b3b4755bb8b10e6d67027.

It should be possible to have multiple applicants with the same email
address and saml_id. Any problems should get identified and addressed
later.

This means we shouldn't be sharing applicants between applications. It
is potentially confusing that updating the application might affect
other applications too. Or might not depending on if we choice an
existing applicant at the start of the process.

Now we always choose a new applicant for every new application, every
time.

Change-Id: Idb119562ae8772f6c4320e6206fc7288c5d27654