Force evaluate get_user value in threadlocals middleware
By default request.user is a SimpleLazyObjectSimple:

request.user = SimpleLazyObject(lambda: get_user(request))

If we do not evaluate request.user in the middleware, request could
change the user's password. The password change is then logged by the
Person.save() method. However the logging tries to retrieve the current
user, which will force evaluation of the request.user value.

This evaluation will call get_user.  However, get_user checks that the
session is still valid, by comparing a hash on the user's password. As
the user's password has just changed, the session is no invalid.
get_user will completely delete the session and log out the user.

The Django code:

def form_valid(self, form):
    session = self.request.session
    user = form.save()
    del self.request.session[INTERNAL_RESET_SESSION_TOKEN]
    if self.post_reset_login:
        auth_login(self.request, user, self.post_reset_login_backend)
    return super().form_valid(form)

When the session is unexpectedly deleted, in the form.save() call,
we crash on the next line where we try to delete a value from the
session which no longer exists.

This in turn causes the test_password_reset_by_self test to fail.

Change-Id: I555b964dfaf0bd5035a7eeb9eb720d48fad858da